requirements.txt
sympy==1.7.1
leidenalg==0.8.8
louvain==0.7.0
matplotlib==3.1.2
numpy==1.18.1
pandas==1.1.5
python-libsbml==5.18.0
scanpy==1.8.2
scipy==1.5.2
pillow==7.0.0
jaal==0.1.2